pyproject.toml
[project]
name = "buildcomputeragent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.96.0",
    "opentelemetry-api>=1.41.1",
    "opentelemetry-exporter-otlp-proto-http>=1.41.1",
    "opentelemetry-sdk>=1.41.1",
]

[tool.uv]
package = true

[tool.setuptools]
package-dir={""="src"}

[tool.setuptools.packages.find]
where=["src"]

[project.scripts]
buildcomputer = "buildcomputer.app:main"
